requirements/tests.txt
coverage
flake8
django-nose
